requirements.txt
pip==22.3.1
asgiref==3.6.0
Django==4.2.7
drf_yasg==1.21.7
djangorestframework==3.14.0
django-rest-authtoken
wheel==0.38.4